pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "github-scout-ai"
version = "0.1.0"
description = "Autonomous AI system that discovers GitHub repos, analyzes them, and generates product ideas"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "Your Name", email = "you@example.com" }
]
keywords = ["github", "ai", "automation", "telegram", "product-ideas"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # Core
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",

    # GitHub API
    "PyGithub>=2.1.0",
    "gql[requests]>=3.5.0",

    # Telegram
    "python-telegram-bot>=20.7",

    # AI/Agents (Phase 2+)
    "crewai>=0.28.0",
    "openai>=1.12.0",

    # Web Framework (Phase 3+)
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",

    # Scheduler
    "apscheduler>=3.10.0",

    # Templates
    "jinja2>=3.1.0",

    # Database
    "aiosqlite>=0.19.0",

    # Utilities
    "httpx>=0.26.0",
    "python-dotenv>=1.0.0",
    "structlog>=24.1.0",
    "tenacity>=8.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
]

[project.scripts]
scout = "src.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100
select = ["E", "F", "W", "I", "N", "UP", "B", "C4"]
ignore = ["E501"]

[tool.ruff.isort]
known-first-party = ["src"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --cov=src --cov-report=term-missing"
